backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# Image Processing
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
imagehash==4.3.1

# Machine Learning
onnxruntime>=1.17.0
torch>=2.1.0
torchvision>=0.16.0
facenet-pytorch>=2.5.3
transformers>=4.36.0  # CLIP for AI tagging

# OCR (optional)
pytesseract==0.3.10

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.1
watchdog==4.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
tqdm==4.66.1
